Fix warning by removing double quotes from test alias in mix.exs
This commit fixes the following warning by remove double quotes
from the test alias in the mix.exs file.

warning: found quoted keyword "test" but the quotes are not
required. Note that keywords are always atoms, even when
quoted, and quotes should only be used to introduce keywords
with foreign characters in them
  /Users/dj.jain/ryan/hapi/deps/queryex/mix.exs:49